+++ src/Storage/Commands.Storage/Blob/Cmdlet/GetAzureStorageContainer.cs
@@ -30,6 +30,7 @@ namespace Microsoft.WindowsAzure.Commands.Storage.Blob.Cmdlet
     /// </summary>
     [Cmdlet(VerbsCommon.Get, StorageNouns.Container, DefaultParameterSetName = NameParameterSet),
         OutputType(typeof(AzureStorageContainer))]
+    [Alias("Get-AzureStorageContainerAcl")]
     public class GetAzureStorageContainerCommand : StorageCloudBlobCmdletBase
     {
         /// <summary>