+++ src/ResourceManager/AzureBatch/ChangeLog.md
@@ -18,6 +18,14 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Set minimum dependency of module to PowerShell 5.0
+* Updated New-AzureBatchPool documentation to remove deprecated example
+
+## Version 4.0.6
+* Updated to the latest version of the Azure ClientRuntime
+
+## Version 4.0.5
+* Fix issue with Default Resource Group in CloudShell
 
 ## Version 4.0.4
 * Added Location Completer to -Location parameters allowing tab completion through valid Locations