+++ src/ResourceManager/RecoveryServices.Backup/ChangeLog.md
@@ -18,6 +18,13 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Set minimum dependency of module to PowerShell 5.0
+
+## Version 4.1.2
+* Updated to the latest version of the Azure ClientRuntime
+
+## Version 4.1.1
+* Fixed issue with cleaning up scripts in build
 
 ## Version 4.1.0
 * Added ResourceGroup Completer to -ResourceGroup parameters allowing tab completion through resource groups in current subscription